requirements/test.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/test.txt requirements/base.in requirements/test.in
#
attrs==17.4.0             # via pytest
coverage==4.5.1           # via pytest-cov
django-model-utils==3.1.1
elasticsearch-dsl==6.1.0
elasticsearch==6.2.0      # via elasticsearch-dsl
funcsigs==1.0.2           # via pytest
ipaddress==1.0.22         # via elasticsearch-dsl
more-itertools==4.1.0     # via pytest
pluggy==0.6.0             # via pytest
py==1.5.3                 # via pytest
pytest-cov==2.5.1
pytest==3.5.1             # via pytest-cov, pytest-django
python-dateutil==2.7.2    # via elasticsearch-dsl
pytz==2018.4              # via django
six==1.11.0               # via elasticsearch-dsl, more-itertools, pytest, python-dateutil
texttable==1.2.1
urllib3==1.22             # via elasticsearch
